Clarify secret_key usages
SECRET_KEY was set to an arbitrary value in certain places in the code but that was only used for build processes. Django would throw an error if it wasn't there during manage.py commands, etc.
Now, it's set to an even more clearly throwaway value. The prod.py SECRET_KEY was always set to pull from ENV variables